requirements.txt
django
djangorestframework
djangorestframework-simplejwt
speechrecognition
openai
python-dateutil
django-cors-headers
python-dotenv
pillow